+++ packages/embeddings/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./ollama.js";
 export * from "./openai.js";
+export * from "./bundled.js";
 export * from "./hash.js";
 export * from "./resolve.js";
 export * from "./enrich.js";